TASK: Use fallback dimension setting if no variant was found
