+++ Classes/EelHelpers/DummyImageSourceHelper.php
@@ -3,24 +3,8 @@ declare(strict_types=1);
 
 namespace Sitegeist\Kaleidoscope\EelHelpers;
 
-use Imagine\Image\Box;
-use Imagine\Image\ImagineInterface;
-use Neos\Flow\Annotations as Flow;
-use Neos\Media\Domain\Model\Adjustment\CropImageAdjustment;
-use Neos\Media\Domain\Model\Adjustment\ImageAdjustmentInterface;
-use Neos\Media\Domain\Model\Adjustment\ResizeImageAdjustment;
-use Neos\Media\Domain\ValueObject\Configuration\Adjustment;
-use Neos\Media\Domain\ValueObject\Configuration\VariantPreset;
-use Neos\Utility\ObjectAccess;
-
 class DummyImageSourceHelper extends AbstractScalableImageSourceHelper
 {
-    /**
-     * @var ImagineInterface
-     * @Flow\Inject
-     */
-    protected $imagineService;
-
     /**
      * @var string
      */
@@ -104,7 +88,9 @@ class DummyImageSourceHelper extends AbstractScalableImageSourceHelper
         $newSource = parent::useVariantPreset($presetIdentifier, $presetVariantName);
 
         if ($newSource->targetImageVariant !== []) {
-            $newSource = $this->calculateSizeFromVariantPresetAdjustments($newSource);
+            $targetBox = $this->estimateDimensionsFromVariantPresetAdjustments($presetIdentifier, $presetVariantName);
+            $newSource->baseWidth = $targetBox->getWidth();
+            $newSource->baseHeight = $targetBox->getHeight();
         }
 
         return $newSource;
@@ -139,70 +125,4 @@ class DummyImageSourceHelper extends AbstractScalableImageSourceHelper
         return $this->baseUri . '?' . http_build_query($arguments);
     }
 
-    /**
-     * @param DummyImageSourceHelper $source
-     * @return ImageSourceHelperInterface
-     * @throws \RuntimeException
-     */
-    protected function calculateSizeFromVariantPresetAdjustments(DummyImageSourceHelper $source): ImageSourceHelperInterface
-    {
-        $presetIdentifier = $source->targetImageVariant['presetIdentifier'];
-        $presetVariantName = $source->targetImageVariant['presetVariantName'];
-
-        $assetVariantPreset = VariantPreset::fromConfiguration($this->variantPresets[$presetIdentifier]);
-        foreach ($assetVariantPreset->variants()[$presetVariantName]->adjustments() as $adjustmentConfiguration) {
-            $adjustment = $this->createAdjustment($adjustmentConfiguration);
-
-            switch (true) {
-                case ($adjustment instanceof ResizeImageAdjustment):
-                    $imageBox = new Box($source->baseWidth, $source->baseHeight);
-                    $image = $this->imagineService->create($imageBox);
-                    if ($adjustment->canBeApplied($image)) {
-                        $image = $adjustment->applyToImage($image);
-                        $source->baseWidth = (int)round($image->getSize()->getWidth());
-                        $source->baseHeight = (int)round($image->getSize()->getHeight());
-                    }
-                break;
-                case ($adjustment instanceof CropImageAdjustment):
-                    $desiredAspectRatio = $adjustment->getAspectRatio();
-                    if ($desiredAspectRatio !== null) {
-                        [, , $newWidth, $newHeight] = CropImageAdjustment::calculateDimensionsByAspectRatio($source->baseWidth, $source->baseHeight, $desiredAspectRatio);
-                    } else {
-                        $newWidth = $adjustment->getWidth();
-                        $newHeight = $adjustment->getHeight();
-                    }
-
-                    $source->baseWidth = (int)round($newWidth);
-                    $source->baseHeight = (int)round($newHeight);
-                break;
-            }
-        }
-
-        return $source;
-    }
-
-    /**
-     * @param Adjustment $adjustmentConfiguration
-     * @return ImageAdjustmentInterface
-     */
-    protected function createAdjustment(Adjustment $adjustmentConfiguration): ImageAdjustmentInterface
-    {
-        $adjustmentClassName = $adjustmentConfiguration->type();
-        if (!class_exists($adjustmentClassName)) {
-            throw new \RuntimeException(sprintf('Unknown image variant adjustment type "%s".', $adjustmentClassName), 1568213194);
-        }
-        $adjustment = new $adjustmentClassName();
-        if (!$adjustment instanceof ImageAdjustmentInterface) {
-            throw new \RuntimeException(sprintf('Image variant adjustment "%s" does not implement "%s".', $adjustmentClassName, ImageAdjustmentInterface::class), 1568213198);
-        }
-        foreach ($adjustmentConfiguration->options() as $key => $value) {
-            ObjectAccess::setProperty($adjustment, $key, $value);
-        }
-
-        if (!$adjustment instanceof ImageAdjustmentInterface) {
-            throw new \RuntimeException(sprintf('Could not apply the %s adjustment to image because it does not implement the ImageAdjustmentInterface.', get_class($adjustment)), 1381400362);
-        }
-
-        return $adjustment;
-    }
 }
